+++ community/kernel/src/test/java/org/neo4j/unsafe/impl/batchimport/cache/idmapping/string/EncodingIdMapperTest.java
@@ -30,7 +30,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
@@ -819,7 +818,7 @@ public class EncodingIdMapperTest
         }
 
         @Override
-        public int badEntries()
+        public long badEntries()
         {
             throw new UnsupportedOperationException();
         }
